pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wiki-cli"
version = "0.1.0"
description = "LLM Wiki - Personal Knowledge Base CLI (Karpathy-style)"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "trafilatura>=1.9",
    "PyYAML>=6.0",
    "python-frontmatter>=1.1",
    "rich>=13.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.25"]
openai = ["openai>=1.0"]
pdf = ["pymupdf4llm>=0.0.5"]
all = ["anthropic>=0.25", "openai>=1.0", "pymupdf4llm>=0.0.5"]

[project.scripts]
wiki = "wiki_cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["wiki_cli*"]
